refactor: extract topo-store-support from topo-support
