+++ apps/trails/src/trails/survey.ts
@@ -23,7 +23,7 @@ import {
   generateSurveyList,
   generateTrailDetail,
 } from './topo-reports.js';
-import { exportCurrentTopo } from './topo-support.js';
+import { exportCurrentTopo } from './topo-store-support.js';
 
 export {
   formatProvisionDetail,
